Edited //sprite to no longer depend on the minidex
//sprite no longer depends on the minidex files from
play.pokemonshowdown.com and instead checks if the female sprite file
exists on the server (using a HTTP HEAD request) should the flag be
provided.

A similar fix (where the sprite's existence is checked) should work for
//furretturretsprite but is not implemented.  As such the minidex files
are still listed as dependencies, though if the command is disabled or
if `commands/dex_commands/furretturretsprite.js` is deleted the minidex
files are no longer required.